// Commons/Isengard/types.ts
export type Stage = "test" | "beta" | "gamma" | "preprod" | "prod";
export type AirportCode =
  | "ALE"
  | "APA"
  | "ARN"
  | "BAH"
  | "BJS"
  | "BOM"
  | "CDG"
  | "CGK"
  | "CMH"
  | "CPT"
  | "DCA"
  | "DUB"
  | "DXB"
  | "FRA"
  | "GRU"
  | "HKG"
  | "HYD"
  | "IAD"
  | "ICN"
  | "KIX"
  | "LCK"
  | "LHR"
  | "LTW"
  | "MEL"
  | "MXP"
  | "NRT"
  | "OSU"
  | "PDT"
  | "PDX"
  | "SFO"
  | "SIN"
  | "SYD"
  | "TLV"
  | "YUL"
  | "ZAZ"
  | "ZHY"
  | "ZRH"
  | "ale"
  | "apa"
  | "arn"
  | "bah"
  | "bjs"
  | "bom"
  | "cdg"
  | "cgk"
  | "cmh"
  | "cpt"
  | "dca"
  | "dub"
  | "dxb"
  | "fra"
  | "gru"
  | "hkg"
  | "hyd"
  | "iad"
  | "icn"
  | "kix"
  | "lck"
  | "lhr"
  | "ltw"
  | "mel"
  | "mxp"
  | "nrt"
  | "osu"
  | "pdt"
  | "pdx"
  | "sfo"
  | "sin"
  | "syd"
  | "tlv"
  | "yul"
  | "zaz"
  | "zhy"
  | "zrh";
export type RegionName =
  | "us-isof-south-1"
  | "us-iso-west-1"
  | "eu-north-1"
  | "me-south-1"
  | "cn-north-1"
  | "ap-south-1"
  | "eu-west-3"
  | "ap-southeast-3"
  | "us-east-2"
  | "af-south-1"
  | "us-iso-east-1"
  | "eu-west-1"
  | "me-central-1"
  | "eu-central-1"
  | "sa-east-1"
  | "ap-east-1"
  | "ap-south-2"
  | "us-east-1"
  | "ap-northeast-2"
  | "ap-northeast-3"
  | "us-isob-east-1"
  | "eu-west-2"
  | "us-isof-east-1"
  | "ap-southeast-4"
  | "eu-south-1"
  | "ap-northeast-1"
  | "us-gov-east-1"
  | "us-gov-west-1"
  | "us-west-2"
  | "us-west-1"
  | "ap-southeast-1"
  | "ap-southeast-2"
  | "me-west-1"
  | "ca-central-1"
  | "eu-south-2"
  | "cn-northwest-1"
  | "eu-central-2";

export type Region = RegionName | AirportCode;

export enum StandardRoles {
  Admin = "Admin",
  FullReadOnly = "FullReadOnly",
  OncallOperator = "OncallOperator",
  ReadOnly = "ReadOnly",
}